requirements.txt
asgiref==3.3.0
Django==3.1.3
django-cors-headers==3.4.0
djangorestframework==3.11.1
pytz==2020.4
sqlparse==0.4.1
gunicorn==20.0.4
psycopg2==2.8.5
whitenoise==5.1.0